+++ types/index.d.ts
@@ -7,7 +7,7 @@
  * before using `validate` middleware.
  */
 declare function init(schemaPath: string, options?: ajvValidatorOptions): void;
-declare function init(jsonSchema: Object, options?: ajvValidatorOptions): void;
+declare function init(jsonSchema: Record<string, any>, options?: ajvValidatorOptions): void;
 export { init };
 
 /**
@@ -18,15 +18,15 @@ export { init };
  * This init variant support loading of external references.
  */
 declare function initAsync(schemaPath: string, options?: ajvValidatorOptions): Promise<void>;
-declare function initAsync(jsonSchema: Object, options?: ajvValidatorOptions): Promise<void>;
+declare function initAsync(jsonSchema: Record<string, any>, options?: ajvValidatorOptions): Promise<void>;
 export { initAsync };
 
 /**
  * Middleware that validates the request against the swagger
  * file, according to the request method and route
  */
-declare function validate(ctx: Object, next: Function): void; // koa
-declare function validate(req: Object, res: Object, next: Function): void; // express
+declare function validate(ctx: Record<string, any>, next: Function): void; // koa
+declare function validate(req: Record<string, any>, res: Record<string, any>, next: Function): void; // express
 declare function validate(options: FastifyPluginOptions): any; // fastify
 export { validate };
 
@@ -56,11 +56,11 @@ export interface FastifyPluginOptions {
 }
 
 export interface ajvValidatorOptions {
-    ajvConfigBody?: object;
-    ajvConfigParams?: object;
+    ajvConfigBody?: Record<string, any>;
+    ajvConfigParams?: Record<string, any>;
     beautifyErrors?: boolean;
     contentTypeValidation?: boolean;
-    errorFormatter?: (errors: ErrorDetails, options: ajvValidatorOptions) => Error;
+    errorFormatter?: (errors: Array<ErrorDetails>, options: ajvValidatorOptions) => Error;
     expectFormFieldsInBody?: boolean;
     firstError?: boolean;
     framework?: frameworks;
@@ -73,3 +73,23 @@ export interface inputValidationOptions {
     beautifyErrors?: boolean;
     firstError?: boolean;
 }
+
+declare class MiddlewareClass {
+    InputValidationError: InputValidationError;
+
+    init(schemaPath: string, options?: ajvValidatorOptions): void;
+    // eslint-disable-next-line no-dupe-class-members
+    init(jsonSchema: Record<string, any>, options?: ajvValidatorOptions): void;
+    initAsync(schemaPath: string, options?: ajvValidatorOptions): Promise<void>;
+    // eslint-disable-next-line no-dupe-class-members
+    initAsync(jsonSchema: Record<string, any>, options?: ajvValidatorOptions): Promise<void>;
+
+    validate(ctx: Object, next: Function): void; // koa
+    // eslint-disable-next-line no-dupe-class-members
+    validate(req: Object, res: Object, next: Function): void; // express
+    // eslint-disable-next-line no-dupe-class-members
+    validate(options: FastifyPluginOptions): any; // fastify
+}
+
+declare function getNewMiddleware(schemaPath: string, options?: ajvValidatorOptions): MiddlewareClass;
+export { getNewMiddleware };