+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Master
 
+### Improvements
+- Update dependencies to address security vulnerabilities
+
+# 3.2.1 - 9 December, 2020
+
+### Fixes
+
+- Add getNewMiddleware to index.d.ts #149
+
 # 3.2.0 - 26 November, 2020
 
 ### Features